\name{Vennerable-package}
\Rdversion{1.1}
\alias{Vennerable-package}
\alias{Vennerable}
\docType{package}
\title{
Venn and Euler area-proportional diagrams
}
\description{
A package for drawing various species of Venn diagrams.
}
\details{
\tabular{ll}{
Package: \tab Vennerable\cr
Version: \tab 1.0\cr
LazyLoad: \tab yes\cr
Depends: \tab graph, RBGL, grid, lattice, RColorBrewer, reshape\cr
License: \tab GPL\cr
Built: \tab R 2.9.0; ; 2009-07-09 13:52:22 UTC; windows\cr
}

Index:
\preformatted{
Venn                    Construct intersections of sets
Venn-class              Class "Venn" ~~~
}

Further information is available in the following vignettes:
\tabular{ll}{
\code{Venn} \tab  (source, pdf)\cr
}
}
\author{
Jonathan Swinton

Maintainer: Jonathan Swinton <jonathan@swintons.net>
}
\keyword{ package }
