\name{VennDiagrams}
\Rdversion{1.1}
\alias{VennDiagrams}
\docType{data}
\title{
Stem cell data
}
\description{
Data on the geometrical structure of various types of Venn diagrams
}
\usage{data(VennDiagrams)}
\format{}
\details{
}
\source{
Created by the \code{buildVennDiagrams()} function within \code{Vennerable}
}
\examples{
data(VennDiagrams)
TD <- VennDiagrams[["ellipses"]][[4]]
class(TD)
}
\keyword{datasets}
